Mise au propre radio-select + renames
