--- src/components/__style__/radio-select.scss
@@ -1,15 +1,0 @@
-.c-radio-select {
-  ul {
-    padding: 0;
-    margin: 0;
-  }
-}
-
-.c-radio-select__title {
-  font-family: "Roboto Condensed";
-  margin-bottom: 4px;
-}
-
-.c-radio-select__option {
-  margin-bottom: 4px;
-}
