+++ src/components/index.ts
@@ -2,7 +2,6 @@ export {default as Autocomplete, AutocompleteProps, AutocompleteStyle} from "./a
 export {default as ButtonBackToTop, ButtonBackToTopProps, ButtonBackToTopStyle} from "./button-back-to-top";
 export {default as ButtonMenu, IconMenu, MenuItem} from "./button-menu";
 export {default as Checkbox, CheckboxProps} from "./checkbox";
-export {default as CheckboxSelect, CheckboxSelectProps, CheckboxSelectStyle} from "./checkbox-select";
 export {default as Display, DisplayProps, DisplayStyle} from "./display";
 export {getIcon} from "./icon";
 export {default as Input, InputProps} from "./input";
@@ -13,4 +12,6 @@ export {default as Panel, PanelButtons, PanelProps, PanelStyle} from "./panel";
 export {default as Popin, PopinStyle} from "./popin";
 export {default as ScrollspyContainer, ScrollspyContainerProps, ScrollspyStyle} from "./scrollspy-container";
 export {default as Select, SelectProps} from "./select";
+export {default as SelectCheckbox, SelectCheckboxProps, SelectCheckboxStyle} from "./select-checkbox";
+export {default as SelectRadio, SelectRadioProps, SelectRadioStyle} from "./select-radio";
 export {default as Switch, SwitchProps} from "./switch";
